Test film counter
