+++ src/index.html
@@ -22,7 +22,7 @@
     </header>
 
     <section class="Main-content">
-      <h2>Shows/Films</h2>
+      <h2 class="films-title">Shows/Films</h2>
       <ul class="shows"></ul>
     </section>
     <section class="popups"></section>
