+++ src/modules/comments.js
@@ -1,16 +1,11 @@
 import {
   getShows,
-  like,
-  getLikes,
   sendCommentApi,
   getCommentsApi,
 } from './api_handler.js';
 
-const shows = document.querySelector('.shows');
 const pops = document.querySelector('.popups');
 const modal = document.querySelector('.modal');
-
-// Comments
 
 const appId = '2lNkwmsdhFTsRqAGHt5J';
 const addComment = async (itemId, username, comment) => {
@@ -90,46 +85,4 @@ const popup = async (id) => {
   document.querySelector('.submit-btn').addEventListener('click', uiAddComment);
 };
 
-const drawFilm = async (id, likes) => {
-  const show = await getShows(id);
-  const data = {
-    item_id: id,
-  };
-  const node = `
-  <li class="show" id="li-${show.id}">
-    <img src="${show.image.medium}" alt="#">
-    <h2>${show.name}</h2>
-    <i class="material-icons like-btn">favorite</i>
-    <span>${likes}</span>
-      <button class="comment-btn" type="button">Comments</button>
-      <button class="reserv-btn" type="button">Reservations</button>
-    </li>`;
-  const child = document.createRange().createContextualFragment(node);
-  shows.appendChild(child);
-
-  shows.querySelector(`#li-${show.id} .comment-btn`).onclick = (e) => {
-    popup(id);
-    pops.classList.toggle('display', (e !== 'popups'));
-    modal.classList.toggle('overlay', (e !== 'modal'));
-  };
-
-  shows.querySelector(`#li-${show.id} .like-btn`).onclick = (e) => {
-    like(data);
-    e.path[0].classList.toggle('liked');
-  };
-};
-
-const render = async () => {
-  const testt = await getShows(7);
-  document.querySelector('.Headline').style.backgroundImage = `url('${testt.image.original}')`;
-
-  const likes = await getLikes();
-  for (let i = 1; i < 15; i += 1) {
-    likes.forEach((item) => {
-      if (i === item.item_id) {
-        drawFilm(i, item.likes);
-      }
-    });
-  }
-};
-render();
+export default popup;
