+++ src/modules/tools.js
@@ -1,0 +1,4 @@
+const counter = (int) => {
+  document.querySelector('.films-title').innerHTML = `Shows/Films (${int})`;
+};
+export default counter;
